# docs/part5-openapi-public-api.yaml
openapi: 3.1.0
info:
  title: SME Sales & Inventory Management API
  version: 1.0.0
  description: >
    Public REST API for the Cloud-Based Sales and Inventory Management
    System with Machine Learning Analytics for SMEs. Served by the
    Express.js backend (Railway Service 1). All endpoints are
    multi-tenant: every authenticated request is implicitly scoped to
    the caller's store_id (BR-01), enforced server-side and never
    accepted as a client-supplied parameter. This service is the only
    public entry point — the FastAPI ML microservice (see the separate
    Internal ML Service specification) is never called directly by
    the frontend.
  contact:
    name: Engineering
  license:
    name: Proprietary — Academic Project
servers:
  - url: https://sme-api.up.railway.app/api/v1
    description: Production (Railway)
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Supabase-issued JWT access token, obtained via POST /auth/login
        or refreshed via POST /auth/refresh. Sent as
        `Authorization: Bearer <token>`. Role (admin/cashier) and
        store_id are resolved server-side from the token's subject —
        never trust client-supplied role/store_id fields.
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
      description: 1-indexed page number.
    LimitParam:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      description: Items per page (BR-21). Max 100.
    SearchParam:
      name: search
      in: query
      schema: { type: string }
      description: Case-insensitive substring search (endpoint-specific field — see each path).
    ProductIdPath:
      name: productId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    DateFromParam:
      name: date_from
      in: query
      schema: { type: string, format: date }
      description: Inclusive range start (UTC date).
    DateToParam:
      name: date_to
      in: query
      schema: { type: string, format: date }
      description: Inclusive range end (UTC date).
  schemas:
    ErrorItem:
      type: object
      properties:
        field: { type: [string, "null"] }
        code: { type: string, example: "INSUFFICIENT_STOCK" }
        message: { type: string }
      required: [code, message]
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }
      required: [page, limit, total, totalPages]
...
    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        store_id: { type: string, format: uuid }
        name: { type: string }
        sku: { type: string }
        category: { type: [string, "null"] }
        price: { type: number, minimum: 0 }
        stock_quantity: { type: integer, minimum: 0 }
        min_stock_threshold: { type: [integer, "null"], minimum: 0 }
        is_low_stock: { type: boolean, description: "Derived: stock_quantity <= effective threshold (BR-11)." }
        image_url: { type: [string, "null"], format: uri }
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      required: [id, store_id, name, sku, price, stock_quantity, is_active, created_at, updated_at]
      examples:
        - id: "b3f1c2a0-1111-4a2b-9c3d-000000000001"
          store_id: "a1b2c3d4-0000-4000-8000-000000000001"
          name: "Kopi Sachet 200g"
          sku: "KOP-200"
          category: "Beverages"
          price: 18500
          stock_quantity: 42
          min_stock_threshold: 10
          is_low_stock: false
          image_url: "https://xyzcompany.supabase.co/storage/v1/object/public/products/kopi-200.jpg"
          is_active: true
          created_at: "2026-05-01T02:00:00Z"
          updated_at: "2026-06-20T08:15:00Z"
    ProductCreateRequest:
      type: object
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        sku: { type: string, minLength: 1, maxLength: 64 }
        category: { type: string, maxLength: 100 }
        price: { type: number, minimum: 0 }
        stock_quantity: { type: integer, minimum: 0, default: 0 }
        min_stock_threshold: { type: integer, minimum: 0 }
      required: [name, sku, price]
      examples:
        - name: "Kopi Sachet 200g"
          sku: "KOP-200"
          category: "Beverages"
          price: 18500
          stock_quantity: 50
          min_stock_threshold: 10
    ProductUpdateRequest:
      type: object
      description: >
        SKU and stock_quantity are intentionally not editable here — stock
        changes must go through POST /inventory/movements (BR-08) so the
        audit log and ML training source remain complete.
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        category: { type: string, maxLength: 100 }
        price: { type: number, minimum: 0 }
        min_stock_threshold: { type: integer, minimum: 0 }
        is_active: { type: boolean }
      minProperties: 1
...
    Transaction:
      type: object
      properties:
        id: { type: string, format: uuid }
        store_id: { type: string, format: uuid }
        cashier_id: { type: string, format: uuid }
        total_amount: { type: number, minimum: 0 }
        created_at: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: '#/components/schemas/TransactionItem' }
      required: [id, store_id, cashier_id, total_amount, created_at, items]
    TransactionItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        product_name: { type: string, description: "Joined at read time for display; not stored on this row." }
        quantity: { type: integer, minimum: 1 }
        unit_price: { type: number, minimum: 0, description: "Price snapshot at time of sale (FR-09)." }
        subtotal: { type: number, minimum: 0 }
      required: [id, product_id, quantity, unit_price, subtotal]
    TransactionCreateRequest:
      type: object
      properties:
        items:
          type: array
          minItems: 1
          items:
            type: object
            properties:
              product_id: { type: string, format: uuid }
              quantity: { type: integer, minimum: 1 }
            required: [product_id, quantity]
      required: [items]
      examples:
        - items:
            - product_id: "b3f1c2a0-1111-4a2b-9c3d-000000000001"
              quantity: 3
            - product_id: "b3f1c2a0-1111-4a2b-9c3d-000000000002"
              quantity: 1
...
  responses:
    BadRequest:
      description: Validation error — malformed or missing fields.
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, const: false }
              message: { type: string, examples: ["Validation failed"] }
              data: { type: "null" }
              errors:
                type: array
                items: { $ref: '#/components/schemas/ErrorItem' }
    Unauthorized:
      description: Missing, malformed, or expired JWT.
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, const: false }
              message: { type: string, examples: ["Unauthorized"] }
              data: { type: "null" }
              errors:
                type: array
                items: { $ref: '#/components/schemas/ErrorItem' }
    Forbidden:
      description: Authenticated, but role or tenant scope does not permit this action.
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, const: false }
              message: { type: string, examples: ["Forbidden - admin role required"] }
              data: { type: "null" }
              errors:
                type: array
                items: { $ref: '#/components/schemas/ErrorItem' }
    NotFound:
      description: Resource not found (or not visible to this tenant).
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, const: false }
              message: { type: string, examples: ["Product not found"] }
              data: { type: "null" }
              errors:
                type: array
                items: { $ref: '#/components/schemas/ErrorItem' }
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List products (Admin and Cashier)
      description: Cashier access is read-only (FR-04); both roles see only their own store's catalog.
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - { $ref: '#/components/parameters/SearchParam' }
        - name: category
          in: query
          schema: { type: string }
        - name: is_active
          in: query
          schema: { type: boolean, default: true }
          description: Filter by active status (BR-22 soft delete).
        - name: sort
          in: query
          schema: { type: string, enum: [name_asc, name_desc, price_asc, price_desc, stock_asc, stock_desc, created_at_desc] }
          description: Default `name_asc`.
      responses:
        "200":
          description: Paginated product list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: true }
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Product' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Products]
      summary: Create a product (Admin only)
      x-required-role: admin
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductCreateRequest' }
      responses:
        "201":
          description: Product created.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: true }
                  message: { type: string, examples: ["Product created"] }
                  data: { $ref: '#/components/schemas/Product' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "409":
          description: SKU already exists for this store (uq_products_store_sku).
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: false }
                  message: { type: string, examples: ["SKU already exists"] }
                  data: { type: "null" }
                  errors:
                    type: array
                    items: { $ref: '#/components/schemas/ErrorItem' }
  /products/{productId}:
    get:
      tags: [Products]
      summary: Get a single product
      parameters: [{ $ref: '#/components/parameters/ProductIdPath' }]
      responses:
        "200":
          description: Product found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: true }
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Product' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Products]
      summary: Update a product (Admin only)
      x-required-role: admin
      parameters: [{ $ref: '#/components/parameters/ProductIdPath' }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductUpdateRequest' }
      responses:
        "200":
          description: Product updated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: true }
                  message: { type: string }
                  data: { $ref: '#/components/schemas/Product' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [Products]
      summary: Soft-delete a product (Admin only)
      description: Sets is_active=false (BR-22). The row is never hard-deleted because products.id is referenced by transaction_items (ON DELETE RESTRICT, Part 3 §3.3).
      x-required-role: admin
      parameters: [{ $ref: '#/components/parameters/ProductIdPath' }]
      responses:
        "200":
          description: Product deactivated.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: true }
                  message: { type: string, examples: ["Product deactivated"] }
                  data:
                    type: object
                    properties:
                      id: { type: string, format: uuid }
                      is_active: { type: boolean, const: false }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "403": { $ref: '#/components/responses/Forbidden' }
        "404": { $ref: '#/components/responses/NotFound' }
...
  /transactions:
    get:
      tags: [Transactions]
      summary: List transactions
      description: Admin sees all transactions for the store; Cashier sees only transactions where cashier_id equals their own user id (BR-06).
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - { $ref: '#/components/parameters/DateFromParam' }
        - { $ref: '#/components/parameters/DateToParam' }
        - name: cashier_id
          in: query
          schema: { type: string, format: uuid }
          description: Admin-only filter; ignored (forced to self) for Cashier callers.
        - name: sort
          in: query
          schema: { type: string, enum: [created_at_asc, created_at_desc], default: created_at_desc }
      responses:
        "200":
          description: Paginated transaction list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: true }
                  message: { type: string }
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Transaction' }
                  meta: { $ref: '#/components/schemas/PaginationMeta' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Transactions]
      summary: Create a POS sale (Cashier or Admin)
      description: >
        Calls `fn_create_sale_transaction` via Supabase RPC as a single
        atomic database transaction (Part 3 §3.6, Part 4 §4.2). Rejects
        the entire sale with 409 if any line item's quantity exceeds
        available stock — no partial sale is ever created (FR-08).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TransactionCreateRequest' }
      responses:
        "201":
          description: Sale completed; stock deducted; receipt data returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: true }
                  message: { type: string, examples: ["Transaction completed"] }
                  data: { $ref: '#/components/schemas/Transaction' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404":
          description: One or more product_id values do not exist for this store.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: false }
                  message: { type: string, examples: ["Product not found"] }
                  data: { type: "null" }
                  errors:
                    type: array
                    items: { $ref: '#/components/schemas/ErrorItem' }
        "409":
          description: Insufficient stock for one or more items — entire transaction rolled back.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, const: false }
                  message: { type: string, examples: ["Insufficient stock for one or more items"] }
                  data: { type: "null" }
                  errors:
                    type: array
                    items: { $ref: '#/components/schemas/ErrorItem' }
...